Add files via upload
This simple code will perform a fuzzy-based prediction using a subtractive clustering scheme. The "Demographic_Prediction_Fuzzy_SC" will extract the desired features which are uploaded as .mat data and process them as .mat files. The sample of the data set is provided with the title "Foot_sex_right". Each process in the code is explained prior to the code syntax.